Cambio en rutas 2.0

+++ routes/api.php
@@ -45,13 +45,13 @@ Route::prefix('auth')->group(function () {
 
 //Rutas Usuarios
 Route::prefix('users')->group(function () {
-    Route::get('list', [UsersController::class, 'index']);
-    Route::get('colaboradores', [UsersController::class, 'colaboradores']);
-    Route::post('create', [UsersController::class, 'store']);
-    Route::get('show/{id}', [UsersController::class, 'show']);
-    Route::put('update', [UsersController::class, 'update']);
-    Route::delete('delete/{id}', [UsersController::class, 'destroy']);
-    Route::get('nombre', [UsersController::class, 'nombre']);
+    Route::get('list', [UsersController::class, 'index'])->middleware('active')->middleware('adminstradores');
+    Route::get('colaboradores', [UsersController::class, 'colaboradores'])->middleware('active')->middleware('roles');
+    Route::post('create', [UsersController::class, 'store'])->middleware('active')->middleware('adminstradores');
+    Route::get('show/{id}', [UsersController::class, 'show'])->middleware('active')->middleware('adminstradores')->where('id', '[0-9]+');
+    Route::put('update', [UsersController::class, 'update'])->middleware('active')->middleware('adminstradores');
+    Route::delete('delete/{id}', [UsersController::class, 'destroy'])->middleware('active')->middleware('adminstradores')->where('id', '[0-9]+');
+    Route::get('nombre', [UsersController::class, 'nombre'])->middleware('active')->middleware('adminstradores');
 });
 
 //Rutas de ideas
@@ -60,7 +60,7 @@ Route::prefix('ideas')->group(function () {
     Route::post('create', [IdeasController::class, 'create']);
     Route::get('userIdeas', [IdeasController::class, 'userIdeas']);
     Route::get('userideasall/{estatus?}', [IdeasController::class, 'userIdeasAll'])->where('estatus', '[0-9]+');
-    Route::get('ideasAll', [IdeasController::class, 'ideasAll'])->where('estatus', '[0-9]+');
+    Route::get('ideasAll/{estatus?}', [IdeasController::class, 'ideasAll'])->where('estatus', '[0-9]+');
     Route::get('show/{id}', [IdeasController::class, 'show'])->where('id', '[0-9]+');
     Route::put('update', [IdeasController::class, 'update']);
     Route::delete('delete/{id}', [IdeasController::class, 'destroy'])->where('id', '[0-9]+');
